requirements.txt
Django==1.11.3
numpy==1.13.0
pkg-resources==0.0.0
pygame==1.9.3
git+https://github.com/alexander-shcheglov/pygameui.git
pytz==2017.2
